[Bug Fix]Use BigDecimal Instead of float
When i start adding really large numbers or divide really small numbers,
the prescion of the answer becomes incorrect.
This is because float is capped to a max and min value.
To fix this i change from type float to BigDecimal instead, as
BigDecimal is capable of dealing with really large numbers and small
numbers.

+++ calculator/src/main/java/entry/SymbolEntry.java
@@ -4,6 +4,7 @@ import enums.EntryInterface;
 import enums.Symbol;
 import enums.Type;
 import exceptions.BadTypeException;
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /**
@@ -40,11 +41,11 @@ public class SymbolEntry implements EntryInterface {
   }
 
   /**
-   * Returns the float value an {@code Entry} represents.
+   * Returns the deciaml value an {@code Entry} represents.
    *
    * @throws BadTypeException If {@code Entry} not of {@code Type NUMBER}
    */
-  public float getValue() {
+  public BigDecimal getValue() {
     throw new BadTypeException("Entry must be Type: NUMBER");
   }
