+++ calculator/src/main/java/model/StandardCalc.java
@@ -2,6 +2,7 @@ package model;
 
 import enums.Symbol;
 import exceptions.InvalidExpressionException;
+import java.math.BigDecimal;
 import stack.OpStack;
 
 /**
@@ -67,7 +68,7 @@ public class StandardCalc implements Calculator {
    * @return The solution for the expression
    */
   @Override
-  public float evaluate(String expression) {
+  public BigDecimal evaluate(String expression) {
     // Split and store components of expression in an array
     String[] expressionParts = expression.split(" ");
     // This regular expression checks if a string is an integer or decimal
